+++ package.json
@@ -45,6 +45,7 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-sort": "^2.4.0",
     "html-webpack-plugin": "^5.5.0",
+    "lodash-webpack-plugin": "^0.11.6",
     "nodemon": "^2.0.15",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.18",
